For Github Action CI, pass action's Python 3 in test call
The prior patch worked in my local environment, but missed a nuance of
Python 3.8 being available in Github's runner, though 3.6 is the
action's configured version.  The Makefile PYTHON3 variable does a
newest-version-first search, so virtualenv wasn't found at runtime.

The name of the action's Python 3 command was found via the home page of
the setup-python action:
https://github.com/marketplace/actions/setup-python

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>