Exclude pandas installation for Python < 3.7
numpy, a pandas dependency, required Python > 3.7 as of version 1.20.0.
Hence, case_sparql_select cannot import pandas in a Python 3.6
environment.

This patch adds steps to disable dependency installation and tests for
case_sparql_select if Python is detected to be < 3.7.x.

As a reminder, Python 3.6 goes EOL on 2021-12-23.  By or at that time,
3.6 will be removed from this project.

References:
* [AC-178] Add website repository's sample runner to
  CASE-Utilities-Python repository
* [PEP-0494] https://www.python.org/dev/peps/pep-0494/#and-beyond-schedule
* [numpy 1.20.0] https://pypi.org/project/numpy/1.20.0/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>