Add mypy source review to current state
This patch is to show what is needed to add static type checking with
`mypy`.

The brunt of the effort is adding `#type: ignore` annotations to
`rdflib`.  These can be removed once an `rdflib` release with the merged
PR 1407 is issued.

The additional package in `tests/requirements.txt` pertaining to
`dateutil` was reported by `mypy`.

DISCLAIMER:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [AC-211] Add static type checking to CASE-Utilities-Python
* https://github.com/RDFLib/rdflib/pull/1407

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>