Add debug-logging option to case_file
This flag was a useful development aid, and its presence in the Make
command lines impacts generated UUID5s.  However, at the end of the
patch series adding this output, all of the debug statements will be
disabled.

A follow-on patch (at the end of the series) will regenerate
Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>