Add copy of monolithic CASE build for a UCO test
UCO Issue 393 added a test that assumed a file's existence during the
UCO CI.  That file is the UCO monolithic build, a purposefully Git-
ignored build artifact.

`case-utils` currently repurposes the UCO `pytest` script to ensure
UCO's tests also pass when substituting the monolithic build of CASE.

With Issue 393, the file reference to the UCO monolithic build needs to
be handled downstream in `case-utils`.  This patch adds that file by
making a copy as a Git-ignored file.  (A soft link could cause potential
confusion due to management of the modification timestamp on the
monolithic build file.)

References:
* https://github.com/ucoProject/UCO/issues/393
* [ONT-295] Release CASE 1.0.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>