Use Warning classes not ignored by default filters
This patch addresses a now-resolved curiosity where
`DeprecationWarning`s were not being visibly raised when intentionally
triggered in an application.  They turn out to be ignored by default.

`UserWarning`s were also not displaying by default, so this patch
designates some of the erroneous behaviors not intended to cause halts
as `RuntimeWarning`s.

This patch also adds the expected environment variable name to a
docstring visible in the command-line help() interface.

References:
* https://docs.python.org/3/library/warnings.html

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>