Update SPARQL query result types
RDFLib 6.3.0, released today, included an update to type hints for the
`Graph.query` function.  Issue 2283 logs some of the new induced type
hints required to get `mypy --strict` to pass.

This patch is needed to respond to the newly induced type assertion
requirements.  It is intentionally based off of `main` in case a
hotfix `case-utils` release is required.

References:
* https://github.com/RDFLib/rdflib/issues/2283

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>