Change start of NamedIndividuals' suffices to use last path-segment of class's IRI
While drafting hand-written example data, it had proved beneficial to
some drafters (myself included) to disambiguate `owl:NamedIndividual`s
from `owl:Class`es by spelling the class differently in the IRI.  Taking
`uco-observable:FileFacet` as an example, it was originally frequently
written as `FileFacet` when referring to the class, and `file-facet-...`
when referring to an individual.

Unfortunately, trying to carry this pattern forward is likely to create
a technological burden.  Camel casing can't always be assumed to apply
straightforwardly, and would cause special-case logic to be needed.
See e.g.:

* `uco-observable:WifiAddressFacet` that would split on capital letters
  to `kb:wifi-address-facet`, which doesn't seem to be a problem;
* `uco-location:GPSCoordinatesFacet` would induce
  `kb:g-p-s-coordinates-facet`, which seems far less obviously
  acceptable;
* `uco-observable:HTTPConnectionFacet` splitting to
  `kb:h-t-t-p-connection-facet` may be the last convincing we need.

Rather than invest in preserving the lowercased, hyphenated suffix
scheme, this patch removes the question and now has individuals use the
last path-segment of the class's IRI.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>