case_validate: Default to None rather than "none"
Defaulting to the string `"none"` has a functional consequence in
downstream pySHACL operations, as encoded at the time of this writing
(pySHACL version 0.23.0).  `inference` in the `pyshacl.validate(...)`
arguments is propagated down two code paths that both impact the
validation operation, but that use different default values.  The
`--metashacl` (SHACL-SHACL validation graph for reviewing shape syntax)-
aligned code path defaults to `rdfs` inference, and the `Validator`
class defaults to the string `"none"`.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>